Add `cursor` kwarg to st.write_stream (#12597)
## Describe your changes

Today, `st.write_stream(...)` appends a ` | ` cursor at the end of the
stream as its being written. This is really nice, but I think we can do
better: let's make the cursor customizable!

This PR introduces the `cursor` kwarg, so when you call
`st.write_stream(..., cursor="🦄")` the final cursor at the end of the
stream is going to be a 🦄. This also supports emoji shortcodes and
material icons.

## GitHub Issue Link (if applicable)

None

## Testing Plan

- ~~Explanation of why no additional tests are needed~~
- [x] Unit Tests (JS and/or Python)
- ~~E2E Tests~~ This is not easy to test E2E since it would depend on
timing.
- ~~Any manual testing needed?~~

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.