[chore] Update dependabot ignore (#12971)
## Describe your changes

- We can not properly* update `react-window` until we also update
`baseui`
- To be specific, we could update our 1 direct usage of `react-window`
in our codebase to v2. However, baseui still has a dependency on
`react-window`, therefore we'd be shipping 2 versions of `react-window`
until we also update baseui. That seems like a strictly worse state than
just being on v1, which works fine.

## GitHub Issue Link (if applicable)

Closes #12621

## Testing Plan

- N/A

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.