Exclude test utilities from production bundle to reduce bundle size (#12966)
## Describe your changes

Some of our test requirements get bundled into our production assets. To
prevent this, we remove the testing stuff from the lib index and instead
provide it as a dedicated import.

An alternative here might be to provide this as an additional package,
but not sure if it is worth the overhead.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>