Enable better file uploading on Android (#13132)
## Describe your changes

Use a higher websocket timeout on Android and better error message to
improve file uploading on Android. Currently, this doesn't work if the
user needs more than 15 seconds to choose a file since the websocket
connection disconnects.

From my research, a proper fix for this would be quite complicate.

I manually verified on Android that this change works and makes it a lot
more convenient to choose files.

## GitHub Issue Link (if applicable)

- Improves https://github.com/streamlit/streamlit/issues/11419 (allows
us to move this from P2 -> P3)

## Testing Plan

- Added unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.