Introduce TransientNode (#12822)
## Describe your changes

Now we will add some transiency. Transient is special, so we will
implement a TransientNode. This change includes adding a method to every
visitor and adding an implementation for the DebugVisitor.

## Testing Plan

- JS Unit Tests
- There's no way this node is introduced yet, so there are no other
tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.