Prevent `st.dialog` from showing elements from previous dialog (#13297)
## Describe your changes

Fixes and issues where elements from previous dialogs were still showing
up in a different dialog as stale elements. This was fixed by always
calculating the dialog ID and only keeping the existing children in case
it's the same dialog -> which expects the same type of children.

Before:


https://github.com/user-attachments/assets/c441b892-d531-4ebe-9178-4bc884b4eeb9

After:


https://github.com/user-attachments/assets/5d8facf9-2592-46e0-b012-b1b5cf3d9ac1


## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/10907

## Testing Plan

- Added e2e and unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.