Add a note about `make update-min-deps`. (#13420)
## Describe your changes

Add a note to `setup.py` about changing min versions of a library.

Without doing this, `py-min-deps-test` will silently use the old min
version.

Note that follow-up work should add a test to verify that this file is
up-to-date, probably as a part of the `py-min-deps-test` action; and
maybe add a precommit to run this automatically.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.